@@ -1,27 +1,35 @@
-/**COMPONENTS */
+/**
+ * COMPONENTS
+ */
 import { AppComponent } from './app.component';
 
 import { RegistrationComponent } from './pages/auth/registration/registration.component';
 import { LoginComponent } from './pages/auth/login/login.component';
 
-/**MODULES */
+/**
+ * MODULES
+ */
 import { NgModule } from '@angular/core';
 import { HttpClientModule } from '@angular/common/http';
 import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { CoreModule  } from "./core/core.module";
+import { CoreModule  } from './core/core.module';
 import { SharedModule } from './shared/shared.module';
 import { MaterialModule } from './shared/material/material.module';
 import { HomeModule } from './pages/home/home.module';
 import { AuthModule } from './pages/auth/auth.module';
+import { BusinessProfileModule } from './pages/business-profile/business-profile.module';
+import { PrivacyModule } from './pages/privacy/privacy.module';
+import { ClientProfileModule } from './pages/client-profile/client-profile.module';
+
 
 
 
 
 @NgModule({
   declarations: [
-    AppComponent,
+    AppComponent
   ],
   imports: [
     BrowserModule,
@@ -32,7 +40,10 @@ import { AuthModule } from './pages/auth/auth.module';
     SharedModule,
     HomeModule,
     AuthModule,
-    HttpClientModule
+    BusinessProfileModule,
+    PrivacyModule,
+    HttpClientModule,
+    ClientProfileModule
   ],
   entryComponents: [RegistrationComponent, LoginComponent],
   providers: [],
