@@ -5,6 +5,7 @@ import {
   FormGroup,
   Validators
 } from '@angular/forms';
+import {MatSnackBar} from '@angular/material/snack-bar';
 
 import { MatDialogRef } from '@angular/material/dialog';
 import { emailRegex } from 'src/app/shared/constants/email';
@@ -18,7 +19,10 @@ export class LoginComponent implements OnInit {
 
   form: FormGroup;
 
-  constructor(private fb: FormBuilder, public matDialogRef: MatDialogRef<LoginComponent>) { }
+  constructor(
+    private _SNACKBAR: MatSnackBar,
+    private fb: FormBuilder,
+    public matDialogRef: MatDialogRef<LoginComponent>) { }
 
   ngOnInit() {
     this.form = this.fb.group({
@@ -29,6 +33,10 @@ export class LoginComponent implements OnInit {
 
   closeClick(): void {
     this.matDialogRef.close();
+    // **HAY QUE SETEAR EL DISEÑO  */
+    this._SNACKBAR.open('Ha ingresado satisfactoriamente', 'ok', {
+      duration: 3000
+    });
   }
 
 }
