@@ -6,7 +6,9 @@ import {
 } from '@angular/forms';
 
 import { MatDialogRef } from '@angular/material/dialog';
-import { ClientService } from 'src/app/services/client/client.service';
+import {MatSnackBar} from '@angular/material/snack-bar';
+
+import { ClientService } from 'src/app/core/services/client.service';
 import { emailRegex } from 'src/app/shared/constants/email';
 import { Client } from 'src/app/shared/interfaces/client';
 
@@ -20,11 +22,13 @@ export class RegistrationComponent implements OnInit {
 
   form!: FormGroup;
 
-  loading = false; //Una vez que se haga submit, loading pasa a  ser verdadero y el boton se deshabilita.
-  
+  loading = false; // Una vez que se haga submit, loading pasa a  ser verdadero y el boton se deshabilita.
+
   // Injeccion de servicios, dialog, formbuilder y servicio cliente.
-  constructor(private fb: FormBuilder, public matDialogRef: MatDialogRef<RegistrationComponent>,
-              private clientService: ClientService) { }
+  constructor(private fb: FormBuilder,
+              public matDialogRef: MatDialogRef<RegistrationComponent>,
+              private clientService: ClientService,
+              private _SNACKBAR: MatSnackBar) { }
 
   // Creación de formgroup.
   ngOnInit() {
@@ -41,44 +45,47 @@ export class RegistrationComponent implements OnInit {
 
   // Getters de cada campo del form.
 
-  get name() { return this.form.get('name') }
+  get name() { return this.form.get('name'); }
 
-  get lastName() { return this.form.get('lastName') }
+  get lastName() { return this.form.get('lastName'); }
 
-  get birthDate() { return this.form.get('birthDate') }
+  get birthDate() { return this.form.get('birthDate'); }
 
-  get email() { return this.form.get('email') } 
+  get email() { return this.form.get('email'); }
 
-  get phoneNumber() { return this.form.get('phoneNumber') }
+  get phoneNumber() { return this.form.get('phoneNumber'); }
 
-  get password() { return this.form.get('password') }
+  get password() { return this.form.get('password'); }
 
-  get passwordConfirm() { return this.form.get('passwordConfirm') }
+  get passwordConfirm() { return this.form.get('passwordConfirm'); }
 
-  get passwordDoesMatch() { return this.password.value === this.passwordConfirm.value }
+  get passwordDoesMatch() { return this.password.value === this.passwordConfirm.value; }
 
   // Implementacion de logica del submit
   async onSubmit() {
-    this.loading = true
+    this.loading = true;
 
     // Objeto cliente.
     // const clientRegister: Client = {
     //   name: this.name.value,
     //   last_name: this.lastName.value,
-    //   birth_date: this.birthDate.value, 
+    //   birth_date: this.birthDate.value,
     //   email: this.email.value,
     //   password: this.password.value
     // }
 
     // const response = await this.clientService.register(clientRegister);
 
-    console.log('loggin test')
+    console.log('loggin test');
   }
 
   // Cierra el modal (Dialog)
   closeClick(): void {
+    this._SNACKBAR.open('Ha ingresado satisfactoriamente', 'ok', {
+      duration: 3000
+    });
     this.matDialogRef.close();
   }
-  
+
 
 }
