@@ -1,4 +1,6 @@
-/**Email Validation */
-export const emailRegex:RegExp = RegExp(
+/**
+ * Email Validation
+ */
+export const emailRegex: RegExp = RegExp(
   /^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/
-  ); 
+);
