@@ -2,8 +2,8 @@ export interface Client {
     name: string;
     last_name: string;
     birth_date: Date;
-    email:string;
-    phone:string;
+    email: string;
+    phone: string;
     password: string;
     state_id: string;
 }
