@@ -1,5 +1,5 @@
-import { Restaurant } from "./restaurant";
-import { Schedule } from "./schedule";
+import { Restaurant } from './restaurant';
+import { Schedule } from './schedule';
 
 export interface Local {
     days: number[];
@@ -16,18 +16,19 @@ export interface Local {
     social_reason: string;
 }
 
-
-/**TODO: MODIFICAR A ESTE PRONTO... */
+/**
+ * TODO: MODIFICAR A ESTE PRONTO...
+ */
 export interface Branch {
     id: number;
     name: string;
-    accept_pet: boolean; 
-    commercial_activity:string;
+    accept_pet: boolean;
+    commercial_activity: string;
     description: string;
-    legal_representative_id:number;   
+    legal_representative_id: number;
     schedule: Schedule;
-    state_id:number;
+    state_id: number;
     restaurant: Restaurant;
-    bank_restaurant_id:number
+    bank_restaurant_id: number;
 
 }
