@@ -1,12 +1,12 @@
-export interface Schedule{
-    id:number;
-    opening_hour:Date;
-    closing_hour:Date;
-    is_monday:boolean;
-    is_tuesday:boolean;
-    is_wednesday:boolean;
-    is_thursday:boolean;
-    is_friday:boolean;
-    is_saturday:boolean;
-    is_sunday:boolean;   
+export interface Schedule {
+    id: number;
+    opening_hour: Date;
+    closing_hour: Date;
+    is_monday: boolean;
+    is_tuesday: boolean;
+    is_wednesday: boolean;
+    is_thursday: boolean;
+    is_friday: boolean;
+    is_saturday: boolean;
+    is_sunday: boolean;
 }
