@@ -1,22 +1,25 @@
 import { NgModule } from '@angular/core';
-/**MODULES ANGULAR */
-import { 
-  MatButtonModule, 
-  MatCardModule, 
-  MatCheckboxModule, 
-  MatDialogModule, 
-  MatDividerModule, 
-  MatFormFieldModule, 
-  MatIconModule, 
-  MatInputModule, 
-  MatListModule, 
-  MatMenuModule, 
+/**
+ * MODULES ANGULAR
+ */
+import {
+  MatButtonModule,
+  MatCardModule,
+  MatCheckboxModule,
+  MatDialogModule,
+  MatDividerModule,
+  MatFormFieldModule,
+  MatIconModule,
+  MatInputModule,
+  MatListModule,
+  MatMenuModule,
   MatPaginatorModule,
   MatRadioModule,
   MatSidenavModule,
   MatSnackBarModule,
-  MatStepperModule, 
-  MatTableModule, 
+  MatStepperModule,
+  MatSelectModule,
+  MatTableModule,
   MatToolbarModule} from '@angular/material';
 import { CdkTableModule } from '@angular/cdk/table';
 
@@ -41,7 +44,8 @@ const MaterialComponents = [
   MatCardModule,
   MatSnackBarModule,
   MatListModule,
-]
+  MatSelectModule
+];
 
 
 @NgModule({
