@@ -1,3 +1,10 @@
 export const environment = {
-  production: true
+  production: true,
+  mapboxToken: 'pk.eyJ1Ijoib3RoZXJ3b3JsZGx5IiwiYSI6ImNrb3hseGkyNzAyOTYyb3RheXk4djh1cDgifQ.OFfTeR-AmNqQGLvKSj7xrA',
+  /**Se deben eliminar cuando las api se encuentren operativas */
+  ms_backoffice: 'https://ms-backoffice-tyne.herokuapp.com/',
+  ms_backbone_clients: 'https://backbone-clients.herokuapp.com/',
+  ms_backbone_reservation: 'https://backbone-reservation.herokuapp.com/',
+  ms_backbone_locals:  '',
+  ms_backbone_menu: 'https://backbone-menu.herokuapp.com/'
 };
