@@ -1,16 +1,11 @@
-// This file can be replaced during build by using the `fileReplacements` array.
-// `ng build --prod` replaces `environment.ts` with `environment.prod.ts`.
-// The list of file replacements can be found in `angular.json`.
-
 export const environment = {
-  production: false
+  production: false,
+  mapboxToken: 'pk.eyJ1Ijoib3RoZXJ3b3JsZGx5IiwiYSI6ImNrb3hseGkyNzAyOTYyb3RheXk4djh1cDgifQ.OFfTeR-AmNqQGLvKSj7xrA',
+  /**Se deben eliminar cuando las api se encuentren operativas */
+  ms_backoffice: 'https://ms-backoffice-tyne.herokuapp.com/',
+  ms_backbone_clients: 'https://backbone-clients.herokuapp.com/',
+  ms_backbone_reservation: 'https://backbone-reservation.herokuapp.com/',
+  ms_backbone_locals:  '',
+  ms_backbone_menu: 'https://backbone-menu.herokuapp.com/'
 };
 
-/*
- * For easier debugging in development mode, you can import the following file
- * to ignore zone related error stack frames such as `zone.run`, `zoneDelegate.invokeTask`.
- *
- * This import should be commented out in production mode because it will have a negative impact
- * on performance if an error is thrown.
- */
-// import 'zone.js/dist/zone-error';  // Included with Angular CLI.
